+++ series/columns_test.go
@@ -7,14 +7,14 @@ import (
 )
 
 func TestSeries_GetColumn(t *testing.T) {
-	newCol := base.Column{Name: "col", Dtype: base.Float}
+	newCol := base.Column{Name: "col", Dtype: base.Float64}
 	col := Column{column: newCol}
 	// assert that Column returned correctly
 	assert.Equal(t, col.GetColumn(), newCol)
 }
 
 func TestSeries_SetColName(t *testing.T) {
-	newCol := base.Column{Name: "col", Dtype: base.Float}
+	newCol := base.Column{Name: "col", Dtype: base.Float64}
 	col := Column{column: newCol}
 	col.SetColName("newColName")
 	// assert that Column name is set correctly
@@ -22,7 +22,7 @@ func TestSeries_SetColName(t *testing.T) {
 }
 
 func TestSeries_SetColIndex(t *testing.T) {
-	newCol := base.Column{Name: "col", Dtype: base.Float}
+	newCol := base.Column{Name: "col", Dtype: base.Float64}
 	col := Column{column: newCol}
 	col.SetColIndex(3)
 	// assert that Column index set correctly
